Move MEG interpolation code interpolation.py
